Пакет storage перенесён в util.storage
