+++ src/main/java/ru/org/linux/site/Template.java
@@ -20,8 +20,8 @@ import org.apache.commons.logging.LogFactory;
 import ru.org.linux.user.auth.AccessViolationException;
 import ru.org.linux.site.config.PathConfig;
 import ru.org.linux.user.*;
-import ru.org.linux.storage.StorageException;
-import ru.org.linux.storage.StorageNotFoundException;
+import ru.org.linux.util.storage.StorageException;
+import ru.org.linux.util.storage.StorageNotFoundException;
 import ru.org.linux.util.LorHttpUtils;
 import ru.org.linux.util.StringUtil;
