+++ src/main/java/ru/org/linux/site/config/StorageConfig.java
@@ -15,7 +15,7 @@
 
 package ru.org.linux.site.config;
 
-import ru.org.linux.storage.Storage;
+import ru.org.linux.util.storage.Storage;
 
 public interface StorageConfig
 {
